+++ sonarqube/CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.4] - 2026-07-17
+
+### 🐛 Bug Fixes
+
+- Bug with src directory default [006f73c](https://github.com/act3-ai/dagger/commit/006f73c33dc4fc65747759f745e433abf15a9aa5) by @pspurlock
+
+
+## [0.0.3] - 2026-07-17
+
+### 🐛 Bug Fixes
+
+- Add --raw-json flag to print json report instead of summary, and --base for image overrides [a68f171](https://github.com/act3-ai/dagger/commit/a68f17156eab84733195962b792036fe12e5bb78) by **Paul Spurlock**
+
+
+## [0.0.2] - 2026-07-16
+
+### 🐛 Bug Fixes
+
+- Add report summary and separate bootstrap function for using GUI [6428c36](https://github.com/act3-ai/dagger/commit/6428c366a3ae6b2638b74d7a1dbb5f6c52d20aec) by @pspurlock
+
+
 ## [0.0.1] - 2026-06-22
 
 ### 🐛 Bug Fixes